requirements.txt
# Core dependencies
python-dotenv>=1.0.0

# Data processing
pandas>=2.0.0
openpyxl>=3.1.0
numpy>=1.24.0
scipy>=1.10.0

# Database
psycopg2-binary>=2.9.0
sqlalchemy>=2.0.0
alembic>=1.13.0

# HTTP (optional - download scripts use stdlib urllib)
requests>=2.31.0
brotli>=1.1.0  # Required for Brotli decompression in requests

# Selenium for browser automation
selenium>=4.15.0